Client-Side: refactor for separation of concerns
