+++ public/javascripts/bishop.js
@@ -1,0 +1,7 @@
+"use strict";
+class Bishop extends Piece {
+    constructor(position) {
+        super(position);
+        this.type = "bishop";
+    }
+}
