+++ public/javascripts/handlers.js
@@ -1,0 +1,7 @@
+"use strict";
+class Handlers {
+    constructor() { }
+    handleStart(event) {
+        console.log("Interesting");
+    }
+}
