+++ public/javascripts/knight.js
@@ -1,0 +1,7 @@
+"use strict";
+class Knight extends Piece {
+    constructor(position) {
+        super(position);
+        this.type = "knight";
+    }
+}
