+++ public/javascripts/queen.js
@@ -1,0 +1,7 @@
+"use strict";
+class Queen extends Piece {
+    constructor(position) {
+        super(position);
+        this.type = "queen";
+    }
+}
