+++ public/javascripts/rook.js
@@ -1,0 +1,8 @@
+"use strict";
+class Rook extends Piece {
+    constructor(position) {
+        super(position);
+        this.type = "rook";
+        this.initial = true;
+    }
+}
