+++ public/stylesheets/styles.css
@@ -1,4 +1,4 @@
-@import url(whitespace-reset.css);
+@import url("./whitespace-reset.css");
 
 html {
   box-sizing: border-box;
