+++ src/@types/javascripts/bishop.d.ts
@@ -1,0 +1,3 @@
+interface Bishop extends GamePiece {
+  type: "bishop";
+}
