+++ src/@types/javascripts/handlers.d.ts
@@ -1,0 +1,3 @@
+interface EventHandlers {
+  handleStart(event: Event): void;
+}
