+++ src/@types/javascripts/king.d.ts
@@ -1,0 +1,6 @@
+interface King extends GamePiece {
+  type: "king";
+  check: boolean;
+  checkMate: boolean;
+  initial: boolean;
+}
