+++ src/@types/javascripts/knight.d.ts
@@ -1,0 +1,3 @@
+interface Knight extends GamePiece {
+  type: "knight";
+}
