+++ src/@types/javascripts/pawn.d.ts
@@ -1,0 +1,7 @@
+interface Pawn extends GamePiece {
+  type: "pawn";
+  initial: boolean;
+  promote(newType: string, player: string, position: string): GamePiece;
+  isPromotableType(type: string): boolean;
+  isPromotablePos(player: string, position: string): boolean;
+}
