+++ src/@types/javascripts/piece.d.ts
@@ -1,0 +1,6 @@
+// represents base piece from which all pieces inherit
+interface GamePiece {
+  position: string | null;
+  setPosition(position: string): string;
+  getPosition(): string;
+}
