+++ src/@types/javascripts/queen.d.ts
@@ -1,0 +1,3 @@
+interface Queen extends GamePiece {
+  type: "queen";
+}
