+++ src/@types/javascripts/rook.d.ts
@@ -1,0 +1,4 @@
+interface Rook extends GamePiece {
+  type: "rook";
+  initial: boolean;
+}
