+++ src/javascripts/handlers.ts
@@ -1,0 +1,7 @@
+class Handlers implements EventHandlers {
+  constructor() {}
+
+  handleStart(event: Event): void {
+      console.log("Interesting");
+  }
+}
