+++ views/layout.pug
@@ -3,8 +3,20 @@ doctype html
 html(lang="en-US")
   head
     title Ajedrez
-    script(type="module" src="javascripts/index.js")
-    script(src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.7.8/handlebars.min.js" integrity="sha512-E1dSFxg+wsfJ4HKjutk/WaCzK7S2wv1POn1RRPGh8ZK+ag9l244Vqxji3r6wgz9YBf6+vhQEYJZpSjqWFPg9gg==" crossorigin="anonymous" referrerpolicy="no-referrer" ) 
+    //- script(type="module" src="javascripts/index.js")
+    script(src="javascripts/gameController.js")
+    script(src="javascripts/board.js")
+    script(src="javascripts/handlers.js")
+    script(src="javascripts/player.js")
+    script(src="javascripts/piece.js")
+    script(src="javascripts/rook.js")
+    script(src="javascripts/queen.js")
+    script(src="javascripts/view.js")
+    script(src="javascripts/pawn.js")
+    script(src="javascripts/bishop.js")
+    script(src="javascripts/king.js")
+    script(src="javascripts/knight.js")
+    //- script(src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.7.8/handlebars.min.js" integrity="sha512-E1dSFxg+wsfJ4HKjutk/WaCzK7S2wv1POn1RRPGh8ZK+ag9l244Vqxji3r6wgz9YBf6+vhQEYJZpSjqWFPg9gg==" crossorigin="anonymous" referrerpolicy="no-referrer" ) 
     link(rel="stylesheet" href="/stylesheets/styles.css")
   body
     header
